back/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
django==3.2
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
django-filter==23.5
    # via -r requirements.in
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.1
    # via -r requirements.in
faker==24.4.0
    # via -r requirements.in
flake8==7.0.0
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
mccabe==0.7.0
    # via flake8
packaging==24.0
    # via
    #   gunicorn
    #   pytest
pillow==10.2.0
    # via -r requirements.in
pluggy==1.4.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
pycodestyle==2.11.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pytest==8.1.1
    # via pytest-django
pytest-django==4.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via faker
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via django
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
uritemplate==4.1.1
    # via drf-spectacular
whitenoise==6.6.0
    # via -r requirements.in
